requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.3
gspread==6.0.0
google-auth==2.27.0
google-generativeai==0.3.2
openpyxl==3.1.2
matplotlib==3.8.2
